Refactoring datatypes

+++ src/Touch.h
@@ -8,8 +8,8 @@
 struct Point
 {
     bool pressed;
-    int x;
-    int y;
+    uint16_t x;
+    uint16_t y;
 };
 
 class Touch
@@ -20,11 +20,11 @@ class Touch
     void waitForTouchReleased();
 
  private:
-    const int MINPRESSURE = 200;
-    const int MAXPRESSURE = 700;
-    const int RESISTANCE  = 350;
-    const int XP = 8, XM = 16, YP = 17, YM = 9; // 240x320 ID=0x9341
-    const int TS_LEFT = 937, TS_RT = 99, TS_TOP = 72, TS_BOT = 915;
+    const uint16_t MINPRESSURE = 200;
+    const uint16_t MAXPRESSURE = 700;
+    const uint16_t RESISTANCE  = 350;
+    const uint8_t XP = 8, XM = 16, YP = 17, YM = 9; // 240x320 ID=0x9341
+    const uint16_t TS_LEFT = 937, TS_RT = 99, TS_TOP = 72, TS_BOT = 915;
 
     TouchScreen ts = TouchScreen(XP, YP, XM, YM, RESISTANCE);
     MCUFRIEND_kbv tft;
